Improved test suite + fixed bug in Site.vectorise()

+++ release.sh
@@ -2,6 +2,12 @@
 echo "Please don't run this as a user. This generates a new release for PyPI. Press ^C to exit or Enter to continue."
 read
 
+if ! python ./testsuite.py && [ "$IGNORE_TEST" != "yes" ]; then
+	echo "The test suite failed. Fix it!"
+	exit 1
+fi
+
+
 # Clear old distutils stuff
 rm -rf build dist MANIFEST &> /dev/null
 # Re-generate site constants
